+++ R/5-rearrange_data.R
@@ -860,7 +860,7 @@ check_caal_data <- function(data_list) {
 
     if (any(is.na(na_check))) {
       na_rows <- which(is.na(na_check))
-      message(sprintf("Composition data have NAs in row(s): %s. Converting to 0s.",
+      message(sprintf("CAAL data have NAs in row(s): %s. Converting to 0s.",
                       paste(na_rows, collapse = ", ")))
       data_list$caal_obs[is.na(data_list$caal_obs)] <- 0
     }